Added getter for _prefix field which allows applications that want to check if the prefix has been set or not. The use case is:
1. Some initialization code configures StatsdClient with MetricsConfig instance
2. Some other part of the application wants to publish the metrics with the now configured client

If the publisher wants to check that the prefix has been set or not, currently there is not way to do that. It can be important to be able to check if prefix was set because it can potentially change the format of the published metric name and thus cause the metrics database to generate improper graphs.

By allowing the getter to retrieve the prefix, the application can intelligently decide whether or not to publish the metric or to configure the correct prefix first.