Stopwatch: introducing "ElapsedTime"
If using a method ElapsedMilliseconds() instead of a property,
(like .NET's Stopwatch has) wasn't enough of a fuck-up*, turns
out that ElapsedMilliseconds() returned int instead of .NET's
Stopwatch's long. So this was making a useless round that could
actually result in hideous bugs because it was even marked as
(unchecked).

Let's then fix it in a prettier way, having a TimeSpan which then
can return the unit of time you wish, instead of suffixing it with
"Milliseconds" (surrendering to PrimitiveObsession).

* because you cannot really fix this problem without breaking
backwards-compatibility (as you cannot have both a method Foo()
and a property Foo in a given class)